Move ReadRepairableExecutionException to the Cassandra package
