+++ core/src/main/java/com/scalar/db/storage/cassandra/ReadRepairableExecutionException.java
@@ -1,4 +1,6 @@
-package com.scalar.db.exception.storage;
+package com.scalar.db.storage.cassandra;
+
+import com.scalar.db.exception.storage.ExecutionException;
 
 public class ReadRepairableExecutionException extends ExecutionException {
